refactor; roughly sort regex by simplicity (#301)
Make sure that many of these can be shortcutted to not invoke
an actual wildcard search at all.
